moved roleTypee from user to production company, now do the same with managerPermission, after what completel delete all role related logic from user
