+++ src/main/java/com/group16b/ApplicationLayer/DTOs/Roles/FounderDTO.java
@@ -1,6 +1,6 @@
 package com.group16b.ApplicationLayer.DTOs.Roles;
 
-import com.group16b.DomainLayer.User.Roles.RoleType;
+import com.group16b.DomainLayer.ProductionCompany.membership.RoleType;
 
 public class FounderDTO extends OwnerDTO{
     public FounderDTO() {
