+++ src/main/java/com/group16b/ApplicationLayer/DTOs/Roles/ManagerDTO.java
@@ -4,7 +4,7 @@ import java.util.Set;
 
 import com.group16b.DomainLayer.User.Roles.Manager;
 import com.group16b.DomainLayer.User.Roles.ManagerPermissions;
-import com.group16b.DomainLayer.User.Roles.RoleType;
+import com.group16b.DomainLayer.ProductionCompany.membership.RoleType;
 
 public class ManagerDTO extends MemberDTO {
     private Integer assignerID;
